Se agrega la funcionalidad de ver producto en la pagina de productohome

+++ src/main/java/com/ecomerce/controller/ProductoController.java
@@ -74,21 +74,21 @@ public class ProductoController {
 	
 	@PostMapping("update")
 	public String update(Producto producto,  @RequestParam("img") MultipartFile file) throws IOException {
+		Producto p = new Producto();
+		p= productoService.get(producto.getId()).get();
+		
 		if(file.isEmpty()) {// opcion para cuando editamos el producto pero no cambiamos la imagen
-			Producto p = new Producto();
-			p= productoService.get(producto.getId()).get();
 			producto.setImagen(p.getImagen());
 		}else {//cuando tambien se edita la imagen
-			Producto p = new Producto();
-			p= productoService.get(producto.getId()).get();
-			
-			if(!p.getImagen().equals("default.jpg"))
-			{
-				upload.deleteImage(p.getImagen());
+			//se deben validar que no se eliman la imagen por defecto
+			if(p.getImagen() != null && !p.getImagen().equals("default.jpg")) {				 
+					upload.deleteImage(p.getImagen());				 
 			}
+			
 			String nombreImagen = upload.saveImage(file);
 			producto.setImagen(nombreImagen);				
 		}
+		producto.setUsuario(p.getUsuario());
 		productoService.update(producto);
 		LOGGER.info("PRODUCTO ACTUALIZADO:{}", producto); 
 		return "redirect:/productos";
